day 73 commit
animator
- rewrote how animations are saved/loaded to save json objects
decomposed into components just
as they are ingame. This makes the loading code of animations ingame a
lot simpler, but in the
animator the saving/loading code is more involved and complicated,
because the data in the
animator isn't layed out the same as ingame. The requirements for the
data are different,
animator needs to be dynamic with easy adding/removing/changing, while
ingame the data is
static, so its optimized for traversal only.